% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Build Response Curves.R
\name{build_response_curves}
\alias{build_response_curves}
\title{Build Response Curves}
\usage{
build_response_curves(
  basedir,
  pa,
  date,
  powers = seq(0.1, 0.9, by = 0.01),
  method = "rlm"
)
}
\arguments{
\item{basedir, pa, date}{Location of directory}

\item{powers}{default from 0.1 to 0.9 by 0.01 - list of powers to test}

\item{method}{rlm is default}
}
\description{
This Function reads in csv's with particular settings
}
\examples{
build_response_curves(basedir,pa,date,powers=seq(0.3,0.9,by=0.01))
}
\keyword{CSV}
\keyword{Read}
