% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write to BQ.R
\name{check_and_write_bq}
\alias{check_and_write_bq}
\title{Write to BQ Function}
\usage{
check_and_write_bq(project_id, dataset, table, data)
}
\arguments{
\item{project_id, dataset, table, data}{Used in checking}

\item{CheckExamples}{for Write to BQ params}
}
\description{
This function write output to BQ after checking to make sure the table exists
}
\examples{
write_to_bq(pa, date, basedir, scenario, table_name, 
dataset = 'interim_solution_forecasting', 
project_id = 'essence-analytics-dwh',
optimization = T,
curves = F,
pa_bq = "")



}
\keyword{curves}
\keyword{plot}
