% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check Planning Inputs.R
\name{check_planning_inputs}
\alias{check_planning_inputs}
\title{Check Planning Inputs}
\usage{
check_planning_inputs(basedir, pa, date, process_data_fun, cost_per_type = 1)
}
\arguments{
\item{basedir, pa, date}{Location of directory}

\item{process_data_fun}{processed data}

\item{cost_per_type}{must be 1 or 2}
}
\description{
Check the planning inputs given
}
\examples{
check_planning_inputs(basedir,pa,date,process_data_fun,cost_per_type=1)
}
\keyword{Check}
\keyword{inputs}
\keyword{planning}
