% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clear for Rerun.R
\name{clear_for_rerun}
\alias{clear_for_rerun}
\title{Clear for Rerun}
\usage{
clear_for_rerun(basedir, pa, date)
}
\arguments{
\item{basedir, pa, date}{Location of directory}
}
\description{
Ensures a rerun is possible
}
\examples{
clear_for_rerun(basedir,pa,date)
}
\keyword{clear_for_rerun}
