% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute Fees.R
\name{compute_fees}
\alias{compute_fees}
\title{Compute Fees}
\usage{
compute_fees(mv_spend, fees)
}
\arguments{
\item{mv_spend, fees}{spend and fees}
}
\description{
A set of functions that
}
\examples{
compute_fees() among others
}
\keyword{compute}
\keyword{fees}
