% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_all_results.R
\name{Create_all_results}
\alias{Create_all_results}
\title{Create All Results}
\usage{
Create_all_results(df1 = b_mv, df2 = b, df3 = o)
}
\arguments{
\item{df1}{b_mv dataset}

\item{df2}{b dataset}

\item{df3}{o dataset}
}
\description{
This function prepares all optimizations for plotting
}
\examples{
Create_all_results(df1,df2,df3)
}
\keyword{Create}
\keyword{all}
\keyword{results}
