% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create Folder.R
\name{create_folder}
\alias{create_folder}
\title{Create Folder}
\usage{
create_folder(basedir, pa, date)
}
\arguments{
\item{basedir, pa, date}{Location of directory}
}
\description{
Creates folders for optimizations
}
\examples{
create_folder(basedir,pa,date)
}
\keyword{creates}
\keyword{folders}
