% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create Matrix Results.R
\name{Create_Matrix_Results}
\alias{Create_Matrix_Results}
\title{Create Matrix Results}
\usage{
Create_Matrix_Results(df1 = b_mv, df2 = b, df3 = o)
}
\arguments{
\item{df1}{b_mv dataset}

\item{df2}{b dataset}

\item{df3}{o dataset}
}
\description{
This function prepares all optimizations for plotting
}
\examples{
Create_all_results(df1,df2,df3)
}
\keyword{Create}
\keyword{Matrix}
\keyword{results}
