% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create Templates.R
\name{create_templates}
\alias{create_templates}
\title{Create Folder}
\usage{
create_templates(basedir, pa, date, row_order = NULL)
}
\arguments{
\item{basedir, pa, date}{Location of directory}
}
\description{
Creates templates for optimizations
}
\examples{
create_folder(basedir,pa,date)
}
\keyword{creates}
\keyword{templates}
