% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Detect and Remove Outliers.R
\name{detect_and_remove_outliers}
\alias{detect_and_remove_outliers}
\title{Detect and Remove Outliers}
\usage{
detect_and_remove_outliers(dat)
}
\arguments{
\item{dat}{processed data}
}
\description{
This function detects and removes outliers in a separate file
}
\examples{
detect_and_remove_outliers()
}
\keyword{outlier}
\keyword{removal}
