% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping_function.R
\name{Grouping_function}
\alias{Grouping_function}
\title{A Grouping function for Plotting}
\usage{
Grouping_function(y)
}
\arguments{
\item{y}{melted dataset to be grouped for plotting}
}
\description{
A Grouping function for Plotting
}
\examples{
Grouping_function()
}
\keyword{grouping}
