% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard Maximization Function.R
\name{maxim_fun_standard}
\alias{maxim_fun_standard}
\title{Standard Maximization Function}
\usage{
maxim_fun_standard(sp, diag_mat)
}
\arguments{
\item{sp}{spend}

\item{diag_mat}{coefficient and power matrix}
}
\description{
Creates Matrix of Spend to the power * coefficient
}
\examples{
maxim_fun_standard(sp,diag_mat)
}
\keyword{Standard}
\keyword{function}
\keyword{maximization}
