% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimal Mix.R
\name{opt_mix}
\alias{opt_mix}
\title{Optimal Mix}
\usage{
opt_mix(
  diag_mat,
  diag_mat_name,
  cm_mat,
  diag_cm_list,
  fee_list,
  process_data_fun,
  cost_per_type = 1,
  num_periods,
  basis = "CV + Fee",
  gran_mult,
  B_max = 1e+08,
  store_inc = 1000,
  precision = 10,
  out_fun = out_fun_standard,
  maxim_fun = maxim_fun_standard,
  out_fun_args = NULL,
  maxim_fun_args = NULL
)
}
\arguments{
\item{CheckExamples}{this has full parameter list}
}
\description{
This function is used to get optimal mix for other functions
}
\examples{
opt_mix(diag_mat,diag_mat_name,cm_mat,diag_cm_list,fee_list,process_data_fun,cost_per_type=1,num_periods,
basis="CV + Fee",gran_mult,B_max=100000000,store_inc=1000,precision=10,
out_fun=out_fun_standard,maxim_fun=maxim_fun_standard,out_fun_args=NULL,
maxim_fun_args=NULL)
}
\keyword{mix}
\keyword{opt}
