% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard Outcome Function.R
\name{out_fun_standard}
\alias{out_fun_standard}
\title{Standard Outcome Function}
\usage{
out_fun_standard(sp, diag_mat)
}
\arguments{
\item{sp}{outcome}

\item{diag_mat}{coefficient and power matrix}
}
\description{
Creates Matrix of outcome to the power * coefficient
}
\examples{
out_fun_standard(sp,diag_mat)
}
\keyword{Outcome}
\keyword{Standard}
\keyword{function}
