% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Perform Optimization.R
\name{perform_optimization}
\alias{perform_optimization}
\title{Performance Optimization}
\usage{
perform_optimization(
  basedir,
  pa,
  date,
  process_data_fun,
  cost_per_type = 1,
  num_periods,
  basis = "CV + Fee",
  scenarios = NULL,
  B_max = 1e+08,
  store_inc = 1000,
  precision = 10,
  out_fun = out_fun_standard,
  maxim_fun = maxim_fun_standard,
  out_fun_args = NULL,
  maxim_fun_args = NULL
)
}
\arguments{
\item{CheckExamples}{this has full parameter list}
}
\description{
This function runs the performance optimization
}
\examples{
opt_mix(basedir,pa,date,process_data_fun,cost_per_type=1,num_periods,
basis="CV + Fee",scenarios=NULL,B_max=100000000,store_inc=1000,precision=10,
out_fun=out_fun_standard,maxim_fun=maxim_fun_standard,out_fun_args=NULL,
maxim_fun_args=NULL)
}
\keyword{Optimization}
\keyword{Performance}
