% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot Curves.R
\name{plot_curves}
\alias{plot_curves}
\title{A Plot Curves Function}
\usage{
plot_curves(
  basedir,
  pa,
  date,
  ms_name,
  out_name,
  mcp_name,
  process_data_fun,
  cost_per_type = 1
)
}
\arguments{
\item{basedir}{Location of Base Directory}

\item{pa, date, ms_name, out_name, mcp_name, process_data_fun, cost_per_type}{additional parameters}
}
\description{
A Plot Curves Function
}
\examples{
plot_curves(basedir,pa,date,ms_name,out_name,mcp_name,process_data_fun,cost_per_type=1)
}
\keyword{curves}
\keyword{plot}
