% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curves2.R
\name{plot_curves2}
\alias{plot_curves2}
\title{A Modified Plot_Curves Function}
\usage{
plot_curves2(
  basedir,
  pa,
  date,
  ms_name,
  out_name,
  mcp_name,
  process_data_fun,
  cost_per_type = 1
)
}
\arguments{
\item{basedir}{Location of Base Directory}

\item{pa, date, ms_name, out_name, mcp_name, process_data_fun, cost_per_type}{additional parameters}
}
\description{
This function allows you to plot curves even when the original plot curves doesn't work
}
\examples{
plot_curves2()
}
\keyword{curves}
\keyword{plot}
