% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process Data DSM.R
\name{process_data_dsm}
\alias{process_data_dsm}
\title{Process Data Function for EMEA Daily Sales}
\usage{
process_data_dsm(
  basedir,
  pa,
  date,
  process = T,
  attr = "LC",
  file = "Data/Raw/DSM_EMEA_Daily Sales Data_working_Q4_2020_to_Q4_2021.csv"
)
}
\arguments{
\item{basedir, pa, date}{Location of directory}

\item{attr}{= 'LC' or MTA}

\item{file}{defaults to DSM_EMEA_Daily Sales Data_working_Q4_2020_to_Q4_2021.csv"}

\item{process_data_fun}{processed data}
}
\description{
This function process provided raw data
}
\examples{
process_data_dsm(basedir,pa,date,process=T, attr = 'LC', file ="Data/Raw/DSM_EMEA_Daily Sales Data_working_Q4_2020_to_Q4_2021.csv")
}
\keyword{Create}
\keyword{all}
\keyword{results}
