% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read CSV.R
\name{read_csv}
\alias{read_csv}
\title{Read CSV Function}
\usage{
read_csv(file, ...)
}
\arguments{
\item{file}{Location of Base Directory}
}
\description{
This Function reads in csv's with particular settings
}
\examples{
read_csv()
}
\keyword{CSV}
\keyword{Read}
